IG-12276: Support pseudo-URL not only in tsdbctl.
